+++ sources/reports/v3/CHANGELOG.md
@@ -1,6 +1,99 @@
 # Sample Schema changes
 ---
 
+
+## Call Event Report
+
+### Revision 1
+ * Created
+
+
+## Client Data Channel Report
+
+### Revision 2
+ * Renamed report from pc-data-channel to client-data-channel
+
+## Call Extension Report
+
+### Revision 1
+ * Created
+
+
+## Client Transport Report
+
+### Revision 2
+ * Renamed report from pc-transport to client-transport 
+
 ## Revision 1
- * Initiated
- 
+ * Created
+
+
+## Inbound Audio Track Report
+
+### Revision 1
+ * Created
+
+
+## Inbound Video Track Report
+
+### Revision 1
+ * Created
+
+## Observer Event Report
+
+### Revision 1
+ * Created
+
+
+## Outbound Audio Track Report
+
+### Revision 1
+ * Created
+
+
+## Outbound Video Track Report
+
+### Revision 1
+ * Created
+
+
+## Report
+
+### Revision 1
+ * Created
+
+
+## Sfu Event Report
+
+### Revision 1
+ * Created
+
+
+## Sfu Inbound Rtp Stream Report
+
+### Revision 1
+ * Created
+
+
+## Sfu Meta Report
+
+### Revision 1
+ * Created
+
+
+## Sfu Outbound Rtp Stream Report
+
+### Revision 1
+ * Created
+
+
+## Sfu SCTP Stream Report
+
+### Revision 1
+ * Created
+
+
+## Sfu Transport Report
+
+### Revision 1
+ * Created